Use `DEBUG` to log `invalid authorization header` in `OAuth2TokenExtractor`
Motivation:

Invalid authorization is a client error so it isn't much a server
maintainer can do. It seems better to use `DEBUG` to detect problems
during development. `WARN` could be noisy because it may be related to
a monitoring system.

Modifications:

- Use `DEBUG` to log an invalid authorization header.

Result:

An invalid authorization header is now logged at `DEBUG` by the OAuth2
token extractor.