Fix a bug where the parent's timeout is not cancelled in `RetryingClient`
Motivation:

Originally, when a request is derived, the parent's timeout scheduler is
not canceled. This was not a problem until the `ResponseTimeoutMode.FROM_START`
feature was added, because the timeout did not start at this point. If `ResponseTimeoutMode.FROM_START`
is set, a timeout task is scheduled before it is derived so the uncanceled
scheduler task may cause a leak because referenced objects are not GC'd
until the task is completed.

Modifications:

- Cancel the timeout scheduler of the parent request when it is derived.

Result:

Fixed a bug where timeout tasks leak when using `ResponseTimeoutMode.FROM_START`
with `RetryingClient`.