`ServerHttpResponse` sets the response `statusCode` correctly (#5884)
Motivation:

It seems like Armeria's spring integration response `ServerHttpResponse`
is not always set correctly.
In detail, it seems like we can confirm that `getStatusCode` returns
`UNKNOWN` unless `setStatusCode` is called explicitly.

1. When a successful response is returned, we refer to the response
headers used. This is similar to the upstream implementation which falls
back to the response which is actually returned.

https://github.com/spring-projects/spring-framework/blob/d2ea5b444812b4c55e00fecb7b6451073677061d/spring-web/src/main/java/org/springframework/http/server/reactive/ReactorServerHttpResponse.java#L72-L76

2. When `Armeria` returns a response instead of `Spring`, the response
is never committed. Now, the `doOnCancel` is listened to and the
response is committed accordingly.

Modifications:

- `ArmeriaServerHttpResponse#getStatusCode` now also checks the response
headers when returning a value
- `doOnCancel` now also tries to set the response code and commit the
`ArmeriaServerHttpResponse`

Result:

- https://github.com/line/armeria/issues/5882

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->