Add StreamMessage.timeout() (#5761)
Motivation:
Currently, the `aggregate()` and `subscribe()` methods of
`StreamMessage` do not have the ability to set a timeout. This provides
the ability to detect when a client or server has not responded for a
period of time and handle it appropriately. Additionally, the timeout
API can be used to detect idle streams by setting a timeout until the
next message.

Modifications:
1. Added the `TimeoutStreamMessage` class
-  Wrap a `StreamMessage` to provide timeout functionality
- You can set a timeout feature by adding the `timeout()` method to the
`StreamMessage` interface.

2. Added the `TimeoutSubscriber` class
- Schedule a timeout for each message.
- `StreamTimeoutMode` allows you to set different timeout modes.

3. Added `StreamTimeoutMode` enumeration
- Defines the `UNTIL_FIRST`, `UNTIL_NEXT`, and `UNTIL_EOS` modes.

4. Added a timeout method
- Added the `timeout` method to the `StreamMessage`, `HttpResponse`, and
`HttpRequest` interfaces to provide the ability to set a timeout.

Result:
- Closes #5744
- This change adds timeout functionality to the `aggregate()` and
`subscribe()` methods of `StreamMessage` and HTTP requests/responses.
- This allows idle streams to be detected and handled appropriately.
- You can use `StreamTimeoutMode` to set the timeout between the arrival
of the first message, the arrival of the next message, or the end of the
stream.

---------

Co-authored-by: Ikhun Um <ih.pert@gmail.com>
Co-authored-by: jrhee17 <guins_j@guins.org>