Fill the stacktrace of `ResponseCompleteException` when `verboseResponses = true`
Motivation:

`ResponseCompleteException` is considered a safe exception to clean up
request resources. However, it would be a good idea to leave a
stacktrace in case there is a bug in the implementation or the user
wants to know how the error occurred.

Modifications:

- Create a new instance if `verboseResponse = true` and
  `ResponseCompletionException` is sampled by `verboseExceptionSampler`

Result:

You can now sees the stacktrace of `ResponseCompleteException` when
`Flags.verboseResponse()` is true.