Exclude protobuf 4.x from `armeria-grpc` module (#5992)
Motivation:

With the recent release of 1.31.0, we received a report that protobuf 4
has been included as an api dependency.
This is probably a mistake since 1) the community isn't ready for
protobuf 4 2) it's usually safer to follow the protobuf version used by
`grpc-java`.

The `api` configuration is a consumable configuration, which means it is
difficult to detect these kind of issues directly.
However, by checking the runtime dependencies for tests we can infer
whether a version is inadvertently bumped.

In order to detect such mishaps, I also propose that a
`failOnVersionConflict` variant is added.
Because naively introducing `failOnVersionConflict` introduces many
conflicts, I've added a variant which checks for specified dependencies
only. (inspired by https://github.com/gradle/gradle/issues/8813)

One limitation is that the `dependencies` task is not available when a
`failOnVersionConflict` occurs. For this reason, once a failure due to a
conflict occurs, it is encouraged to use the `-PdebugDeps` flag with the
`dependencies` task.

e.g.
```
./gradlew :grpc:dependencies -PdebugDeps
```

Modifications:

- Excluded the `protobuf-java` dependency from `protobuf-jackson`
- Added a `failOnVersionConflict` method which checks version conflicts
for specific dependencies only

Result:

- Closes #5990 

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->