Fill the stack trace of `ResponseCompleteException` when sampled (#5972)
Motivation:

`ResponseCompleteException` is considered a safe exception for cleaning
up request resources. However, it would be a good idea to leave a stack
trace in case there is a bug in the implementation or the user wants to
know how the error occurred.

Discord thread:
https://discord.com/channels/1087271586832318494/1087272728177942629/1303562249629073520

Modifications:

- Create a new instance if `ResponseCompletionException` is sampled by
`verboseExceptionSampler`

Result:

You can now sample the stack trace of `ResponseCompleteException` with
`verboseExceptionSampler`.