Fix a bug where stale endpoints in `KubernetesEndpointGroup` aren't updated
Motivation:

When a Kubernetes service is updated, a new Pod watcher is created and
receives events. At this point, the old cache should have been cleared,
but it was not cleared, so stale endpoints were exposed.

Modifications:

- Clear `podToNode` map when a new Pod watcher is created

Result:

Fixed a bug where stale endpoints remained when a Kubernetes service was updated.