Fix a bug where stale endpoints in `KubernetesEndpointGroup` aren't updated (#6012)
Motivation:

When a Kubernetes service is updated, a new Pod watcher is created and receives events. The old cache should have been cleared at this point, but it was not removed, so stale endpoints were exposed.

Modifications:

- Clear `podToNode` map when a new Pod watcher is created

Result:

Fixed a bug where stale endpoints remained when a Kubernetes service was updated.