Periodically remove inactive connection pool metrics
Motivation:

We observed that threads were blocked when multiple connections were
closed simultaneously and the endpoint has a small number of event
loops.
https://github.com/line/armeria/blob/fa76e99fa6132545df3a8d05eeb81c5681ec8953/core/src/main/java/com/linecorp/armeria/client/ConnectionPoolMetrics.java#L79-L85

We have no exact evidence, but I guess the `remove()` operation of
Micrometer may take a long time. The other logic is a simple HashMap
operation, so it does not block for a long time.

Modifications:

- Add a dedicated gc thread to remove inactive meters whose active
  conections is 0.
  - A jitter is added to prevent gc from executing at the same time.
  - Unsed meters are removed every an hour + jitter.

Result:

- Fix the bug where `EventLoop` is blocked for a long time by
  `ConnectionPoolListener.metricCollecting()` when a connection is closed.