Stop updating endpoints when `HealthCheckedEndpointGroup` is closing
Motivation:

When `HealthCheckedEndpointGroup` is closing, it clears the internal
endpoints.
https://github.com/line/armeria/blob/011741dc51cf328b78488655e78ceca1cdc0a1b3/core/src/main/java/com/linecorp/armeria/client/endpoint/healthcheck/DefaultHealthCheckerContext.java#L135
In fact, the closing process is not a problem. However, when the
endpoints changes, the change is propagated to `EndpointSelector`.

Some `EndpointSelector` implementations warn about the empty endpoints.
https://github.com/line/armeria/blob/62da203bd9e11d669476e6999937352749cd7339/core/src/main/java/com/linecorp/armeria/client/endpoint/WeightedRoundRobinStrategy.java#L81
The warning is not a real problem but a false positive alarm.

For better compatibility with `EndpointSelector` implementations, I
propose to stop additional endpoints from being updated when
`HealthCheckedEndpointGroup` is closed. As a result, `EndpointSelector`
will not updated with empty endpoints.

Modifications:

- Stop updating endpoints when `HealthCheckedEndpointGroup` is closing

Result:

You no loger see a false positive warning log when
`HealthCheckedEndpointGroup` is closing.