Fix intermittent brave test failure (#6093)
Motivation:

Currently, `BraveClientIntegrationTest` is an abstract class which is
shared by two tests. However, when tests are run in parallel it is
possible that the shared `ClientFactory` is closed after executing a
test class. This will result in failure for the other test class
extending `BraveClientIntegrationTest`.



```
./gradlew test

...
Http1BraveClientIntegrationTest > reportsClientKindToZipkin() FAILED
    java.lang.IllegalStateException: Cannot create a client because the factory is closing.
        at app//com.linecorp.armeria.client.DefaultClientFactory.newClient(DefaultClientFactory.java:171)
        at app//com.linecorp.armeria.client.AbstractWebClientBuilder.buildWebClient(AbstractWebClientBuilder.java:152)
...
```

ref: https://ge.armeria.dev/s/ea3vdvjs6utag

Modifications:

- It seems like even if the default client factory is used, the test
passes. Modified to just use the default client factory.

Result:

- Brave tests pass reliably

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->