Support java modules for thrift (#6076)
Motivation:

We received a report that thrift modules do not support java 9 modules.
When checking the specification, it seems like a identifier needs to
start with a JavaLetter.
ref:
https://docs.oracle.com/javase/specs/jls/se16/html/jls-7.html#jls-7.4

Unfortunately, our thrift modules contain a dot followed by a digit,
which isn't a JavaLetter.
i.e. `thrift0.9` where 9 isn't a digit.

I propose that for thrift modules, the last dot is omitted to avoid this
issue.
i.e. `thrift0.9` -> `thrift09`

Modifications:

- Introduced a `automaticModuleNameOverride` so that the override
configuration can live inside each module
- `automaticModuleName` is now null if not configured, or a provider if
configured. This change is necessary so that the actual value of
`automaticModuleName` is not determined at configuration time, giving
each module a chance to configure its own `ext` property.
- Each Jar task now configures `Automatic-Module-Name` at execution time
as opposed to configuration time. This is also necessary so that each
module can configure its own `ext` property prior to setting the name.
- Each thrift module now sets its own `automaticModuleNameOverride`

Result:

- Closes https://github.com/line/armeria/issues/6075

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->