Initialize a Docusaurus project for the new Armeria site (#6098)
Motivation:

We're preparing for the migration of the Armeria website to reduce
maintenance costs. The new website is built with Docusaurus, which
offers features and components specifically designed for documentation.
This approach alleviates the need for us to develop and maintain the
website on our own.

Modifications:

- Initialize a Docusaurus project
- Configure some initial settings
- Set up eslint and prettier
- Set up antd and less

Result:

- The new project is ready for development.
- Note that it has some peer dependency issues, which are expected to be
resolved in the future Docusaurus releases (e.g.
https://github.com/facebook/docusaurus/issues/10817)