fixes customizer logic in the GsonGrpcJsonMarshallerBuilder
JsonFormat.Parser and JsonFormat.Printer classes are immutable and
calling any method in customizer creates new object.
Because of that result of the customer is ignored so we need a way
to use new object created in the customizer.
Replaces Consumer<T> with the Function<T, T> customizer which
returns customized object.