Allow configuring the maximum age of Kubernetes `Watch` (#6117)
Motivation:

Currently, the KubernetesEndpointGroup watch is designed to stay active
for as long as possible and automatically restart when an error occurs.
However, in practice, maintaining a long-running watch session
inevitably leads to errors, causing it to terminate and restart. From
the perspective of monitoring observers, these errors could be
false-positive noise.

To help prevent these errors, I would suggest periodically closing the
watch when it reaches a certain lifespan and creating a new one.

Modifications:

- Add `maxWatchAge()` option to `KubernetesEndpointGroupBuilder`
  - If unset, the maximum age of a `Watch` defaults to 10 minutes.
- Refactor `KubernetesEndpointGroup` to start a new watch service when
the maximum age is reached.

Result:

You can now set the maximum age of a `Watch` used in
`KubernetesEndpointGroup`.