Remove DnsCacheListener from DnsCache when resolver is closed (#6174)
Motivation:

The `RefreshingAddressResolver` is added to the `DnsCache` as listener
but is never removed from it when the resolver is closed. This makes the
`DnsCache` hold a strong reference (internally, in the
`CopyOnWriteArrayList` that stores all the listeners) of it even when
the resolver is closed, which may cause memory leak if the `DnsCache` is
a global one and lives for the lifetime of the JVM.


Modifications:

- Removes the `DnsCacheListener` from the `DnsCache` to make it eligible
for GC.

Result:

- Closes https://github.com/line/armeria/issues/6173.
- When the resolver is closed, the `DnsCache` no longer stores it in the
list and hence no longer holds a reference of it.