Handle a race condition where `onSubscribe` may be emitted downstream after `onError` (#6168)
Motivation:

While debugging
https://github.com/line/armeria/issues/5880#issuecomment-2735723709, I
found that `onSubscribe` could be emitted after `onError`.
This is a clear violation of specifications
https://github.com/reactive-streams/reactive-streams-jvm?tab=readme-ov-file#1.7
and
https://github.com/reactive-streams/reactive-streams-jvm?tab=readme-ov-file#1.9.

I propose that before aborting, `onSubscribe` is called if it hasn't
been called yet. To prevent `onSubscribe` being called multiple times,
`invokedOnSubscribe` is used as a guard.

Modifications:

- Extracted `maybeInvokeOnSubscribe` and called it also from
`notifySubscriberOfCloseEvent0`

Result:

- We no longer observe race conditions between `onSubscribe` and
`onError` when aborting a `StreamMessage`.