Use `ctx.authority()` for logging unexception client exceptions
Motivation:

If an `EndpointGroup` is used to send the request, `request.authority()`
will be null. Consequently, log messages for unexpected exceptions lack
the remote peer information which makes it harder users to identity
the root cause.

Modification:

- Use `ctx.authority()` in the warning mesage.

Result:

Remote peer information is correctly logged when an unexpected request
failure occurs on the client side.