Ensure `RequestLog` completion when the response is fully processed (#5714)
Motivation:

Related: #5708, #4834

If a decorator directly returns a response, the `RequestLog` for the
request won't be complete. Because the `RequestLog` is automatically
completed when it starts to be written on the wire by
`AbstractRequestHandler`.

In the case of `RetryingClient`, `RequestLog`'s events are used to get
exceptions to apply `RetryRule`. If the `RequestLog` events aren't
complete, `RetryingClient` will fall into a deadlock.

Typically, `RequestLog` is automatically completed so it is difficult
for the user to recognize that `RequestLog` must be completed.
Consequently, it is easy to miss.

I propose two changes in this PR to solve the problem.

1. If `RequestLog.requestFirstBytesTransferredTimeNanos` isn't logged
when a response completes, we can assume that the response is returned
by a decorator. A hook is added to a response completion event to
complete the log manually for that case.

2. `RetryingClient` is refactored to manually set some properties of
`RequestLog` by using `SplitHttpResponse` or `AggregatedHttpResponse`
when the `ResponseLog` isn't logged by `AbstractRequestHandler`.

Modifications:

- Add a hook that completes `RequestLog` if it is written to wire to
`ClientUtil`
- Split `handleResponse` into `handleAggregatedResponse` and
`handleStreamingResponse` in `RetryingClient`
- The headers of `SplitHttpResponse` is used to set
`RequestLog.responseHeaders()` if the RequsetLog is incomplete.
- If `SplitHttpResponse.headers()` completes exceptionally, the cause is
set to `RequestLog.responseCause()` and ends `RequestLog` immediately.

Result:

- Fixed deadlock that could occur in `RetryingClient` due to
`RequestLog` not being completed.
- Fixed a bug where `RequestLog` is not completed when the HTTP response
is fully consumed on the client.