Introduce `Preprocessor` based clients (#6060)
Motivation:

#6057 has described preprocessors, which allow users to dynamically
modify how to connect to a target endpoint dynamically. However, there
is a limitation where an `EndpointGroup` and `SessionProtocol` must be
provided even though the `Preprocessor` will solely determine how to
connect to a remote.

For example, specifying the protocol and endpoint when building the
client is not necessary since the preprocessor will fill in all of the
necessary information.
```java
HttpPreprocessor preprocessor = (delegate, ctx, req) -> {
  ctx.setEndpoint(Endpoint.of(...));
  ctx.setSessionProtocol(HTTPS);
  return delegate.execute(ctx, req);
}
GrpcClients.of(HTTP, Endpoint.of(...)).preprocessor(preprocessor)
```

I propose that users are able to build a client solely based on
`Preprocessor`. By doing so, we may later introduce a `XdsPreprocessor`
which may dynamically determine how to connect to a remote based on
resources.

```java
GrpcClients.of(XdsPreprocessor.of(...));
```

In order to support the above, I propose the following:
- If a client is built based on `Preprocessor`, the `SessionProtocol`
must be `UNDEFINED`.
- For instance,
`GrpcClients.of(HttpPreprocessor).params().scheme().sessionProtocol()`
will be `UNDEFINED`
- If a client is built with `SessionProtocol.UNDEFINED`, at least one
`Preprocessor` must be present in `ClientOptions`.
- For instance, a client built via
(WebClient.of("undefined://google.com")) must also supply a
preprocessor.
- An exception to this rule is `WebClient.of()`. Because it is easy to
supply a target `EndpointGroup`, `SessionProtocol` from its API, a
preprocessor is not necessarily for creating a default WebClient.

Modifications:

- User facing APIs have been added which allow clients to be initialized
solely based on a `Preprocessor`.
- Users may specify `Preprocessor`, and optionally a SerializationFormat
and prefix path.

With the addition preprocessor-based clients, it is important to
validate whether a client was created incorrectly. In order to validate
client parameters from a unified location, it is necessary to know how a
client was created. For this reason, following changes have been made to
ClientBuilderParams:
- If a client is not created based on a URI, create an internal URI
based on a rule so that we can check how a client was created. e.g.
endpointGroup based clients will have prefix `armeria-group-`,
preprocessor based clients will have prefix `armeria-preprocessor-`,
etc...
- `ClientBuilderParamsUtil` has utility methods which determines how a
client has been created.
- Introduce a `ClientBuilderParamsBuilder` which introduces a fluent way
to create a new `ClientBuilderParams` based on an existing
`ClientBuilderParams`. This will help preserve the URI authority which
embeds how a client was created.

Result:

- Users can create clients solely based on `Preprocessor`

Miscellaneous

Just because an exception has been thrown from preprocessing does not
necessarily mean that the request has not been sent. I have removed the
`UnprocessedRequestException` wrapping behavior.
e.g.
```
val preprocessor = (delegate, ctx, req) -> {
  val res = delegate.execute(ctx, req);
  throw new RuntimeException();
}
```

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->