Support multi-value query parameters (#6118)
## Motivation:

This PR is a follow-up to
[#6072](https://github.com/line/armeria/pull/6072), which added support
for mapping all query parameters into a Map<String, String>. However, it
does not handle cases where multiple values exist for the same query
parameter key.

For example:
```
GET /foo?a=1&a=2&b=3
```
The current implementation only retains the last value `a=2`, losing
`a=1`.
To better support multi-value query parameters, we introduce support for
Map<String, List<?>>.

## Modifications:

- Enhanced ofQueryParamMap in AnnotatedValueResolver to support:
  - Map<String, String> → Single-value mapping (existing behavior).
  - Map<String, List<String>> → Multi-value mapping (new behavior).
- If the value type of the map is List<?>, query parameters with the
same key are collected into a list.
- Added validation to ensure that the value type is List<?>. If an
unsupported type is used, an IllegalArgumentException is thrown.

## Result:

- Users can now use @Param Map<String, List<?>> to collect multi-value
query parameters.

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->

Signed-off-by: Daeho Kwon <trewq231@naver.com>