Leave request/response content for `AnnotatedService`s (#6231)
Motivation:

Before introducing log masking for annotated services, it would be
useful if annotated services record `RequestLogBuilder#requestContent`
and `RequestLogBuilder#responseContent` out of the box.

This PR attempts to record the request parameters and the response value
by default.
The request parameters and return values are wrapped in an
`AnnotatedRequest` or `AnnotatedResponse` as suggested in the following
[comment](https://github.com/line/armeria/issues/5711#issuecomment-2461464935).

It should also be noted that parameters and values are set as-is without
post-processing. When logging, it is the responsibility of the logger to
unwrap values such as `CompletableFuture` or `Optional`. As users using
`JsonLogFormatter` with `AnnotatedService` may be affected by this
change, the `Flags#jsonContentLogging` flag is introduced to turn off
the feature entirely.

Now that all major services (gRPC/thrift/rest) now all record `*Content`
by default. On the other hand, `ContentPreviewingService` is an opt-in
feature where the user must explicitly add a decorator.
Users will not have a way to view the actually outgoing/incoming logs
using `ContentPreviewing*` since `[request|response]Content` is almost
always set. In order to handle this, `ContentPreviewing*` is now
recorded to `preview`, whereas `[request|response]Content` is recorded
to `content`.

Lastly, using `JsonLogFormatter` naively applies `ObjectMapper` to
request parameters and return values. In order to reduce the possibility
of encountering an exception while logging, `FAIL_ON_EMPTY_BEANS` as
well as `AnnotatedRequestJsonSerializer` and
`AnnotatedResponseJsonSerializer` are introduced.

Retrofit/Feign unfortunately do not provide extension points for setting
the request/response content reliably. Hence, the client-side corollary
will be dealt with separately.

Modifications:

- Introduced `Flags#jsonContentLogging` to give an option to turn off
json content logging globally
- Introduced `AnnotatedRequest` and `AnnotatedResponse` which wrap the
parameters and return values when logging request/response content
- Modified `DefaultAnnotatedService#invoke` to record request
parameters/return values

Result:

- Closes #5711

<!--
Visit this URL to learn more about how to write a pull request
description:

https://armeria.dev/community/developer-guide#how-to-write-pull-request-description
-->